build: fix x32 compiler warning for printf formats
common/ECLogger.cpp: In member function 'size_t KC::Logger::MakeTimestamp(char*, size_t)':
common/ECLogger.cpp:144:62: warning: format '%ld' expects argument of type 'long int', but argument 4 has type '__syscall_slong_t' {aka 'long long int'} [-Wformat=]
  144 |  snprintf(buffer + strlen(buffer), z - strlen(buffer), ".%06ld", ts.tv_nsec / 1000);
ECtools/stats/stats.cpp: In function 'std::string mapitable_ToString(const SPropValue*)':
ECtools/stats/stats.cpp:465:35: warning: format '%lx' expects argument of type 'long unsigned int', but argument 4 has type 'long long int' [-Wformat=]
  465 |   snprintf(buf, sizeof(buf), "0x%lx", lpProp->Value.li.QuadPart);
libfreebusy/recurrence.cpp: In member function 'bool KC::recurrence::CheckAddValidOccr(time_t, time_t, time_t, const KC::TIMEZONE_STRUCT&, ULONG, KC::OccrInfo**, ULONG*)':
libfreebusy/recurrence.cpp:951:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Wformat=]
  951 |  ec_log_debug("Testing match: %lu ==> %s", tsNow, ctime(&tsNow));
libfreebusy/recurrence.cpp:953:16: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Wformat=]
libfreebusy/recurrence.cpp:958:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
libfreebusy/recurrence.cpp:986:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Wformat=]
libfreebusy/recurrence.cpp:987:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Wformat=]
libfreebusy/recurrence.cpp:988:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
libfreebusy/recurrence.cpp:989:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
libfreebusy/recurrence.cpp:1159:17: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
libfreebusy/recurrence.cpp:1168:17: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
libfreebusy/recurrence.cpp:1178:16: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Wformat=]
ECtools/monitor/ECQuotaMonitor.cpp:96:15: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'long long int' [-Wformat=]
provider/server/ECThreadManager.cpp:144:30: warning: format '%ld' expects argument of type 'long int', but argument 11 has type 'long long int' [-Wformat=]
provider/server/ECThreadManager.cpp:145:32: warning: format '%ld' expects argument of type 'long int', but argument 14 has type 'long long int' [-Wformat=]
provider/server/ECThreadManager.cpp:146:32: warning: format '%ld' expects argument of type 'long int', but argument 17 has type 'long long int' [-Wformat=]
php-ext/main.cpp:219:33: warning: format '%lu' expects argument of type 'long unsigned int', but argument 6 has type 'std::chrono::duration<long long int, std::ratio<1, 1000> >::rep' {aka 'long long int'} [-Wformat=]
provider/plugins/LDAPUserPlugin.cpp: In member function 'void LDAPUserPlugin::my_ldap_search_s(const char*, int, const char*, const char* const*, int, LDAPMessage**, LDAPControl**)':
provider/plugins/plugin.h:36:49: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'LONGLONG' {aka 'long long int'} [-Wformat=]
provider/plugins/LDAPUserPlugin.cpp:655:2: note: in expansion of macro 'LOG_PLUGIN_DEBUG'

+++ ECtools/monitor/ECQuotaMonitor.cpp
@@ -93,7 +93,9 @@ void* ECQuotaMonitor::Create(void* lpVoid)
 	if(hr != hrSuccess)
 		kc_perror("Quota monitor failed", hr);
 	else
-		ec_log_info("Quota monitor done in %lu seconds. Processed: %u, Failed: %u", tmEnd - tmStart, lpecQuotaMonitor->m_ulProcessed, lpecQuotaMonitor->m_ulFailed);
+		ec_log_info("Quota monitor done in %lld seconds. Processed: %u, Failed: %u",
+			static_cast<long long>(tmEnd - tmStart),
+			lpecQuotaMonitor->m_ulProcessed, lpecQuotaMonitor->m_ulFailed);
 	return NULL;
 }
